LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-optgroup-arrow-keys.optional-expected.txt


one
two
three
four
five
six

FAIL Keyboard navigation forwards and backwards should visit each option with optgroups. assert_equals: First option should be initially focused. expected Element node <option class="one">one</option> but got Element node <select>
  <optgroup label="group1">
    <option class="o...
